Replace Globals with PathUtils and SettingsLoader
