+++ bin/SyncDBData.rb
@@ -50,8 +50,8 @@ display_help if $opts.has_key?('--help')
 # ------------------------------------
 
 # Settings
-require 'Globals.rb'
-Globals.load_settings(ARGV.size > 0 ? ARGV[0] : nil)
+require 'SettingsLoader.rb'
+SettingsLoader.load_settings(ARGV.size > 0 ? ARGV[0] : nil)
 
 # Logger
 require 'BuildLogger.rb'
